requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.10.7
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.129
llvmlite==0.39.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy==1.1.1
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@b5851c6c40e753606765ac45b85b298e3ae9e00d
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==12.1.3
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==1.4.47
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
